added cp4d catalg sources

+++ .github/scripts/validate-deploy.sh
@@ -66,5 +66,23 @@ if [[ $count -eq 20 ]]; then
   exit 1
 fi
 
+
+
+count=0
+until kubectl get CatalogSource "ibm-cpd-ccs-operator-catalog" -n "openshift-marketplace" || [[ $count -eq 20 ]]; do
+  echo "Waiting for subscription/cpd-operator in openshift-marketplace"
+  count=$((count + 1))
+  sleep 15
+done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for CatalogSource/ibm-cpd-ccs-operator-catalog in openshift-marketplace"
+  kubectl get all -n "openshift-marketplace"
+  exit 1
+fi
+
+
+
+
 cd ..
 rm -rf .testrepo
